Rename package "enums" to "util"
