+++ src/main/java/de/adesso/util/Command.java
@@ -1,4 +1,4 @@
-package de.adesso.enums;
+package de.adesso.util;
 
 public enum Command {
